+++ qml/TransactionDelegate.qml
@@ -117,8 +117,15 @@ ItemDelegate {
             font.styleName: 'Medium'
             text: {
                 if (transaction.amounts.length === 1) {
-                    wallet.displayUnit
-                    return transaction.amounts[0].formatAmount(true)
+                    const amount = transaction.amounts[0]
+                    if (transaction.type === Transaction.Outgoing) {
+                        const network = transaction.account.wallet.network
+                        if (!network.liquid || amount.asset.isLBTC) {
+                            const key = network.liquid ? network.policyAsset : 'btc'
+                            return formatAmount(transaction.data.satoshi[key])
+                        }
+                    }
+                    return amount.formatAmount(true)
                 } else {
                     return qsTrId('id_multiple_assets')
                 }